Sprite sheet icon system, disable Matrix UI, documentation
Major Changes:
- Replaced vector icons with AI-generated sprite sheets (20+ sprite sheets)
- Created sprite sheet loader classes for all icon categories
- Added icon alignment fix for upgrade tab clipping
- Disabled Matrix prestige feature (UI hidden, code preserved)
- Removed rank display from EPS area
- Added comprehensive CLAUDE.md documentation for AI continuity

New Files:
- Utils/*SpriteSheetLoader*.cs (18 new loaders)
- Resources/*.png (20+ sprite sheets)
- CLAUDE.md, DEVELOPMENT.md, BALANCE_TODO.md

Icon System:
- IconHelper.CreateIconWithFallback() uses 18-step fallback chain
- Each sprite sheet loader maps icon IDs to (row, col) positions
- Supports dynamic icon sizing with padding

+++ Models/Quest.cs
@@ -14,7 +14,7 @@ public class Quest
     public required string Description { get; init; }
     public required string FlavorText { get; init; }
     public required QuestRisk Risk { get; init; }
-    public required int BelieversRequired { get; init; }
+    public required double BelieversRequired { get; init; }
     public required int DurationSeconds { get; init; }
     public required double SuccessChance { get; init; } // 0.0 to 1.0
     public string Icon { get; init; } = "?";
@@ -22,7 +22,7 @@ public class Quest
     // Rewards on success
     public double EvidenceReward { get; init; }
     public double EvidenceMultiplier { get; init; } = 1.0; // Multiplied by current EPS
-    public int TinfoilReward { get; init; }
+    public long TinfoilReward { get; init; }
     public string? UnlockUpgradeId { get; init; }
 
     // Partial reward on fail (for Low risk)
@@ -34,7 +34,7 @@ public class ActiveQuest
     public required string QuestId { get; init; }
     public required DateTime StartTime { get; init; }
     public required DateTime EndTime { get; init; }
-    public required int BelieversSent { get; init; }
+    public required double BelieversSent { get; init; }
 
     public bool IsComplete => DateTime.Now >= EndTime;
     public double Progress => Math.Min(1.0, (DateTime.Now - StartTime).TotalSeconds / (EndTime - StartTime).TotalSeconds);
